requirements.txt
pandas
numpy
datetime
google.oauth2
time
asyncio
gspread
tabula-py
requests
